requirements.txt
streamlit
PyYAML
langchain
openai
langchain-openai
pydantic
langchain-community
python-dotenv
python-docx
pypdf
tiktoken
pdfkit
markdown
yfinance
plotly
pandas
yahooquery
